engine/requirements.txt
# ── Core Framework ──
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Database ──
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# ── Cache & Message Queue ──
redis>=5.0.0
celery>=5.4.0

# ── Market Data ──
yfinance>=0.2.38
pandas>=2.2.0
numpy>=1.26.0

# ── Technical Analysis ──
ta>=0.11.0

# ── HTTP Client ──
httpx>=0.27.0
websockets>=12.0

# ── Auth & Security ──
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# ── Utilities ──
python-dotenv>=1.0.1
structlog>=24.1.0
pyyaml>=6.0.1
jsonschema>=4.22.0
croniter>=2.0.5

# ── Testing ──
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0
